[Sprite Lab] unwind customBlocks/spritelabBlocks
