+++ apps/src/block_utils.js
@@ -1,5 +1,11 @@
 import _ from 'lodash';
 
+import {
+  updatePointerBlockImage,
+  updatePointerBlockWarning,
+} from '@cdo/apps/blockly/addons/cdoSpritePointer';
+import {spriteLabPointers} from '@cdo/apps/p5lab/spritelab/blockly/constants';
+
 import {BlockColors, BlockStyles, EMPTY_OPTION} from './blockly/constants';
 import cdoBlockStyles from './blockly/themes/cdoBlockStyles';
 import MetricsReporter from './metrics/MetricsReporter';
@@ -947,7 +953,44 @@ exports.createJsWrapperBlockCreator = function (
             );
         }
 
-        Blockly.customBlocks.setUpBlockShadowing.bind(this)();
+        // We only set up block shadowing for blocks that have a type in spriteLabPointers.
+        if (Object.keys(spriteLabPointers).includes(this.type)) {
+          // saveExtraState is used to serialize the image source block ID.
+          this.saveExtraState = function () {
+            return {
+              imageSourceId: this.imageSourceId,
+            };
+          };
+
+          // loadExtraState is used to deserialize the image source block ID.
+          // We use this id to set the initial pointer block image.
+          this.loadExtraState = function (state) {
+            this.imageSourceId = state['imageSourceId'];
+            if (this.imageSourceId) {
+              updatePointerBlockImage(
+                this,
+                spriteLabPointers,
+                this.imageSourceId
+              );
+              const imageSourceBlock = Blockly.getMainWorkspace()?.getBlockById(
+                this.imageSourceId
+              );
+              if (imageSourceBlock) {
+                const imageSourceBlockWorkspace = imageSourceBlock.workspace;
+                imageSourceBlockWorkspace.addChangeListener(event => {
+                  onBlockImageSourceChange(event, this);
+                });
+              }
+            }
+          };
+
+          // When the block's parent workspace changes, we check to see if
+          // we need to update the shadowed block image or warning text.
+          this.onchange = function (event) {
+            onBlockImageSourceChange(event, this);
+            updatePointerBlockWarning(this, spriteLabPointers);
+          };
+        }
 
         interpolateInputs(blockly, this, inputRows, inputTypes, inline);
         this.setInputsInline(inline);
@@ -1119,3 +1162,38 @@ const getHelpUrl = function (docFunc) {
   // Documentation is only available for Sprite Lab.
   return `/docs/spritelab/${docFunc}`;
 };
+
+// On change event for a block that shadows an image source block.
+// On an event, checks if the block image should change, and update it.
+function onBlockImageSourceChange(event, block) {
+  const imagePreview =
+    block.inputList && block.inputList[0] && block.inputList[0].fieldRow[1];
+  if (!imagePreview) {
+    return;
+  }
+  if (event.type === Blockly.Events.BLOCK_DRAG && event.blockId === block.id) {
+    // If this is a start event, prevent image changes.
+    // If it is an end event, allow image changes again.
+    imagePreview.setAllowImageChange(!event.isStart);
+  }
+  if (
+    (event.type === Blockly.Events.BLOCK_CREATE &&
+      event.blockId === block.id) ||
+    (event.type === Blockly.Events.BLOCK_CHANGE && event.blockId === block.id)
+  ) {
+    // We can skip the following events:
+    // This block's create event, as we handle setting the image on block creation
+    // in src/p5lab/spritelab/blocks.
+    // This block's change event, as that means we just changed the image due to
+    // some other event.
+    return;
+  }
+  if (
+    imagePreview.shouldAllowImageChange() &&
+    (event.type === Blockly.Events.BLOCK_CREATE ||
+      event.type === Blockly.Events.BLOCK_CHANGE ||
+      event.type === Blockly.Events.BLOCK_DRAG)
+  ) {
+    updatePointerBlockImage(block, spriteLabPointers);
+  }
+}
