+++ apps/src/blockly/customBlocks/cdoBlockly/index.js
@@ -2,10 +2,8 @@
 
 import {blocks as commonBlocks} from './commonBlocks';
 import {blocks as playlabBlocks} from './playlabBlocks';
-import {blocks as spritelabBlocks} from './spritelabBlocks';
 
 export default {
   ...commonBlocks,
   ...playlabBlocks,
-  ...spritelabBlocks,
 };
