+++ apps/src/blockly/customBlocks/googleBlockly/index.ts
@@ -2,10 +2,8 @@
 
 import {blocks as commonBlocks} from './commonBlocks';
 import {blocks as playlabBlocks} from './playlabBlocks';
-import {blocks as spritelabBlocks} from './spritelabBlocks';
 
 export default {
   ...commonBlocks,
   ...playlabBlocks,
-  ...spritelabBlocks,
 };
