+++ apps/src/p5lab/spritelab/blocks.js
@@ -1,8 +1,16 @@
-import {NO_OPTIONS_MESSAGE} from '@cdo/apps/blockly/constants';
+import {Order} from 'blockly/javascript';
+
+import {BLOCK_TYPES, NO_OPTIONS_MESSAGE} from '@cdo/apps/blockly/constants';
+import {blocks as behaviorBlocks} from '@cdo/apps/blockly/customBlocks/googleBlockly/behaviorBlocks';
+import {
+  editButtonHandler,
+  toolboxConfigurationSupportsEditButton,
+} from '@cdo/apps/blockly/customBlocks/googleBlockly/proceduresBlocks';
 import {parseSoundPathString} from '@cdo/apps/blockly/utils';
 import {SVG_NS} from '@cdo/apps/constants';
 import {spriteLabPointers} from '@cdo/apps/p5lab/spritelab/blockly/constants';
 import {getStore} from '@cdo/apps/redux';
+import {getAlphanumericId} from '@cdo/apps/utils';
 import i18n from '@cdo/locale';
 import spritelabMsg from '@cdo/spritelab/locale';
 
@@ -287,19 +295,71 @@ const customInputTypes = {
   },
   behaviorPicker: {
     addInput(blockly, block, inputConfig, currentInputRow) {
-      const dropdownField = new Blockly.FieldBehaviorPicker(
-        Blockly.customBlocks.getAllBehaviorOptions
-      );
+      const noBehaviorLabel = i18n.behaviorsNotFound();
+      const noBehaviorOption = [noBehaviorLabel, NO_OPTIONS_MESSAGE];
+      // Behavior definition blocks are always moved to the hidden workspace.
+      const definitionWorkspace = Blockly.getHiddenDefinitionWorkspace();
+      if (!definitionWorkspace) {
+        return [noBehaviorOption];
+      }
+      const behaviorBlocks = definitionWorkspace
+        .getTopBlocks()
+        .filter(block => block.type === BLOCK_TYPES.behaviorDefinition);
+      // Menu options are an array, each option containing a human-readable part,
+      // and a language-neutral string. Both are the same in this case.
+      const behaviorOptions = behaviorBlocks.map(block => [
+        block.getProcedureModel().getName(),
+        block.behaviorId,
+      ]);
+      behaviorOptions.sort();
+      // Add a "No behaviors found" option, if needed
+      if (behaviorOptions.length === 0) {
+        behaviorOptions.push(noBehaviorOption);
+      }
+      const dropdownField = new Blockly.FieldBehaviorPicker(behaviorOptions);
       currentInputRow
         .appendField(inputConfig.label)
         .appendField(dropdownField, inputConfig.name);
-      Blockly.customBlocks.addBehaviorPickerEditButton.call(
-        this,
-        block,
-        inputConfig,
-        currentInputRow,
-        dropdownField
-      );
+      const behaviorsFound =
+        dropdownField.getOptions().length > 1 ||
+        dropdownField.getOptions()[0][1] !== NO_OPTIONS_MESSAGE;
+
+      // Behavior editing is only permitted using the modal function editor.
+      if (
+        behaviorsFound &&
+        Blockly.useModalFunctionEditor &&
+        // TODO: Support editing behaviors from within a modal editor workspace.
+        block.workspace.id === Blockly.getMainWorkspace().id &&
+        toolboxConfigurationSupportsEditButton(block)
+      ) {
+        const editButton = new Blockly.FieldButton({
+          value: i18n.edit(),
+          onClick: editButtonHandler,
+          colorOverrides: {button: 'blue', text: 'white'},
+          allowReadOnlyClick: true, // We support showing the editor even if viewing in read only mode.
+        });
+        block.inputList[block.inputList.length - 1].appendField(
+          editButton,
+          'EDIT'
+        );
+        // getProcedureModel is defined on procedure blocks as part of
+        // @blockly/block-shareable-procedures
+        // For this block, we will get the procedure based on selected
+        // dropdown field option.
+        block.getProcedureModel = function () {
+          const fieldValue = block.getFieldValue(inputConfig.name);
+          const procedureMap = block.workspace.getProcedureMap();
+          let procedure = undefined;
+          for (const value of procedureMap.values()) {
+            if (value.getName() === fieldValue) {
+              procedure = value;
+              break;
+            }
+          }
+          // We should always find the procedure in the map.
+          return procedure;
+        };
+      }
     },
     generateCode(block, arg) {
       const fieldValue = block.getFieldValue(arg.name);
@@ -429,7 +489,122 @@ export default {
       ]
     ));
 
-    Blockly.customBlocks.installBehaviorBlocks(behaviorEditor);
+    Blockly.common.defineBlocks(behaviorBlocks);
+
+    generator.forBlock.behavior_definition = function (_block, generator) {
+      const block = _block;
+      if (!generator.nameDB_) {
+        return null;
+      }
+      // If we don't have a behavior id, generate a random id.
+      // This ensures the hidden definition block will generate valid code.
+      if (!block.behaviorId) {
+        block.behaviorId = getAlphanumericId();
+      }
+      // Define a procedure with a return value.
+      const funcName = generator.nameDB_.getName(
+        block.behaviorId,
+        Blockly.Names.NameType.PROCEDURE
+      );
+
+      // Holds the additional code that is prefixed (injected before every statement) and/or
+      // suffixed (injected after every statement) to the main block of code
+      let xfix1 = '';
+      if (generator.STATEMENT_PREFIX) {
+        xfix1 += generator.injectId(generator.STATEMENT_PREFIX, block);
+      }
+      if (generator.STATEMENT_SUFFIX) {
+        xfix1 += generator.injectId(generator.STATEMENT_SUFFIX, block);
+      }
+      if (xfix1) {
+        xfix1 = generator.prefixLines(xfix1, generator.INDENT);
+      }
+      let loopTrap = '';
+      if (generator.INFINITE_LOOP_TRAP) {
+        loopTrap = generator.prefixLines(
+          generator.injectId(generator.INFINITE_LOOP_TRAP, block),
+          generator.INDENT
+        );
+      }
+
+      // Translate all the inner blocks within the current block into code
+      const branch = generator.statementToCode(block, 'STACK');
+      // Sprite Lab behavior blocks do not have return inputs, but this check is included
+      // in case we'd like to support that in the future.
+      let returnValue =
+        (block.getInput('RETURN') &&
+          generator.valueToCode(block, 'RETURN', Order.NONE)) ||
+        '';
+
+      // Contains the same code as xfix1 if both are present, but applied before the return statement
+      let xfix2 = '';
+      if (branch && returnValue) {
+        xfix2 = xfix1;
+      }
+      if (returnValue) {
+        returnValue = generator.INDENT + 'return ' + returnValue + ';\n';
+      }
+      const args = [];
+      args.push(
+        generator.nameDB_.getName(
+          i18n.thisSprite(),
+          Blockly.Names.NameType.VARIABLE
+        )
+      );
+      const variables = block.getVars();
+      for (let i = 0; i < variables.length; i++) {
+        args[i] = generator.nameDB_.getName(
+          variables[i],
+          Blockly.Names.NameType.VARIABLE
+        );
+      }
+      let code =
+        'function ' +
+        funcName +
+        '(' +
+        args.join(', ') +
+        ') {\n' +
+        xfix1 +
+        loopTrap +
+        branch +
+        xfix2 +
+        returnValue +
+        '}';
+      // Once we are on V11, we can remove this cast as scrub_ will no longer be protected.
+      code = generator.scrub_(block, code);
+      // Add % so as not to collide with helper functions in definitions list.
+      generator.provideFunction_('%' + funcName, code);
+      return null;
+    };
+    generator.forBlock.gamelab_behavior_get = function (_block, generator) {
+      const block = _block;
+      // Generating 'undefined' mimics the code for a missing block.
+      const undefinedCode = ['undefined', Order.ATOMIC];
+      // If we don't have a behavior Id, find on the definition block.
+      if (!this.behaviorId) {
+        const procedureModel = block.getProcedureModel();
+        // If there's no model, fail gracefully.
+        if (!procedureModel) {
+          return undefinedCode;
+        }
+        const definitionBlock = Blockly.Procedures.getDefinition(
+          procedureModel.getName(),
+          Blockly.getHiddenDefinitionWorkspace()
+        );
+        block.behaviorId = definitionBlock?.behaviorId;
+        // If we somehow still don't have a behavior id, fail gracefully.
+        if (!this.behaviorId) {
+          return undefinedCode;
+        }
+      }
+      if (block.behaviorId && generator.nameDB_) {
+        const name = generator.nameDB_.getName(block.behaviorId, 'PROCEDURE');
+        return [`new Behavior(${name}, [])`, Order.ATOMIC];
+      } else {
+        return null;
+      }
+    };
+    generator.forBlock.sprite_parameter_get = generator.forBlock.variables_get;
     Blockly.Blocks.sprite_variables_get = {
       // Variable getter.
       init: function () {
